Update README.md
When the water-level is below the minimum detectable level (MDL), the seven segment display is arranged to show the digit 0, indicating that the tank is empty, when the water reaches level1 (but is below level2) the connection between the probes gets completed (through the conducting medium – water) and the base voltage of transistor increases.

This causes the base-emitter junction of transistor to get forward biased, this switches transistor from cut-off to conduction mode thus PIN (B7) of microcontroller is pulled to ground hence, the corresponding digit displayed by the seven segment display is 1.

The similar mechanism applies to the detection of all the other levels. When the tank is full, all input pins of microcontroller become low. This causes the display to show 8 and also in this case a buzzer sound is given, thereby indicating a “tank full” condition.

Most water level indicators are equipped to indicate and detect only a single level. The Water Level Indicator implemented here can indicate up to nine such levels and the microcontroller displays the level number on a seven segment display.
So, the circuit not only capable of cautioning a person that the water tank has been filled up to certain level, but also indicates that the water level has fallen below the minimum detectable level. This circuit is important in appliances such as the water cooler where there is a danger of motor-burnout when there is no water in the radiator used up also it can be used in fuel level indication.

Recommened Reading: Water Level Alarm Using 555 Timer

In this project we show the water level indicator using eight transistors which conducts as level rises, a buzzer is also added which will automatically start as the water level becomes full, auto buzzer start with the help of microcontroller. With the help of this project we not only show the level of water on seven segment display but also indicate the water full condition using a buzzer.
Water Level Indicator Project Circuit Features:
Easy installation.
Low maintenance.
Compact elegant design.
The Automatic water level controller ensures no overflows or dry running of pump there by saves electricity and water.
Avoid seepage of roofs and walls due to overflowing tanks.
Fully automatic, saves man power.
Consume very little energy, ideal for continuous operation.
Automatic water level controller provides you the flexibility to decide for yourself the water levels for operations of pump set.
Shows clear indication of water levels in the overhead tank.
How Water Level Indicator Project Circuit Works?
The operation of this project is very simple and can be understood easily. In our project “water level indicator” there are 3 main conditions:

There is no water available in the source tank.
Intermediate level i.e. either of 3rd to 7th level.
There is ample amount of water available in the source tank.
So let us discuss more about these 3 conditions

CONDITION 1: Water not available
When the tank is empty there is no conductive path between any of the 8 indicating probes and the common probe (which is connected to 5v+ supply) so the transistor base emitter region will not have sufficient biasing voltage hence it remains in cut off region and the output across its collector will be Vc approximately 4.2v.

As in this case the microcontroller is used in the active low region (which means it considers 0-2 volts for HIGH and 3-5 volts for LOW) now the output of transistor which is 4.2v approximately will be considered as LOW by the microcontroller and hence the default value given by microcontroller to the seven segment display is 0 which indicates as the tank is empty.

CONDITION 2: Intermediate levels
Now as the water starts filling in the tank a conductive path is established between the sensing probes and the common probe and the corresponding transistors get sufficient biasing at their base, they starts conducting and now the outputs will be Vce (i.e. 1.2v-1.8v) approximately which is given to microcontroller.

Here the microcontroller is programmed as a priority encoder which detects the highest priority input and displays corresponding water level in the seven segment display.

In this project while the water level reaches the 7th level i.e. last but one level along with display in seven segment a discontinuous buzzer is activated which warns user that tank is going to be full soon.

CONDITION 3: Water full
When the tank becomes full, the top level probe gets the conductive path through water and the corresponding transistor gets into conduction whose output given to microcontroller with this input microcontroller not only displays the level in seven segment display but also activates the continuous buzzer by which user can understand that tank is full and can switch off the motor and save water.
Water Level Indicator Project Applications:
Automatic Water level Controller can be used in Hotels, Factories, Homes Apartments, Commercial Complexes, Drainage, etc., It can be fixed for single phase motor, Single Phase Submersibles, Three Phase motors. (For 3Æ and Single Phase Submersible Starter is necessary) and open well, Bore well and Sump. We can control two motor and two sumps and two overhead tanks by single unit.
Automatic water level controller will automatically START the pump set as soon as the water level falls below the predetermined level (usually 1/2 tank) and shall SWITCH OFF the pump set as soon as tank is full.
Fuel level indicator in vehicles.
Liquid level indicator in the huge containers in the companies.